// frontend/src/types/fileResponse.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FileResponseStatus } from './fileResponseStatus';

export interface FileResponse {
  id?: number;
  originalName?: string;
  objectName?: string;
  status?: FileResponseStatus;
}